requirements.txt
asgiref==3.10.0
click==8.3.0
colorama==0.4.6
Django==5.2.7
django-unfold==0.67.0
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
Pillow==10.1.0
psycopg2==2.9.6
psycopg2-binary==2.9.5
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.37.0
whitenoise==6.11.0
